requirements.txt
streamlit
pandas
requests
python-dotenv
transformers
plotly
wordcloud
matplotlib